dashboard/requirements.txt
streamlit
tensorflow-macos
tensorflow-metal
numpy
Pillow

